Comments on Intro
Some grammatical changes. Removed unnecessary clauses. Specified who the artifacts should be available to (one of our assigned projects might have been available within the company it was produced at, but wasn't available to us)

Use the phrase "In this paper" a little repetitively, including in three consecutive sentences - tried to re-word some of them.


Also, 6626 seems like a REALLY exact number. I'm assuming that's taken from the estimates we gave at the end of the course? I know mine was an estimate. Would it be better to round down and say over 6600 hours?

Related to previous - I'm assuming not everyone has responded? Can we still say "We" spent that time considering that some of the hours were not put in by the authors (updated to "we and our classmates").